% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fl_colorscales.R
\name{cs_blue}
\alias{cs_blue}
\title{Create a colour palette of blues.}
\usage{
cs_blue(n)
}
\arguments{
\item{n}{The number of colours of blue to return.}
}
\value{
A character vector of hexadecimal values.
}
\description{
This function will create a lovely, evenly spaced colour palette.
}
\examples{
cs_blue(10)

}
\author{
Bernard Gentili
}
