% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fl_colorscales.R
\name{cs_blye}
\alias{cs_blye}
\title{Create a colour palette from blues to yellows.}
\usage{
cs_blye(n)
}
\arguments{
\item{n}{The number of colours of blue to yellow to return.}
}
\value{
A character vector of hexadecimal values.
}
\description{
This function will create a lovely, evenly spaced colour palette.
}
\examples{
cs_blye(10)

}
\author{
Bernard Gentili
}
