% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fl_colorscales.R
\name{cs_BuYlRd}
\alias{cs_BuYlRd}
\title{Create a colour palette from blues to reds.}
\usage{
cs_BuYlRd(n)
}
\arguments{
\item{n}{The number of colours of blue to red to return.}
}
\value{
A character vector of hexadecimal values.
}
\description{
This function will create a lovely, evenly spaced colour palette.
}
\examples{
cs_BuYlRd(10)

}
\author{
Bernard Gentili
}
