% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fl_DownloadFjord.R
\name{fl_DownloadFjord}
\alias{fl_DownloadFjord}
\title{Calculate yearly means for event metrics.}
\usage{
fl_DownloadFjord(fjord, dirdata = "FjordLight.d")
}
\arguments{
\item{fjord}{Expects a character vector for one of the 8 available fjords.}

\item{dirdata}{The directory where the user would like to download the data.
Default is "FjordLight.d".}
}
\value{
The downloaded NetCDF file contains the following variables:
  \item{bathymetry}{depth [m]}
  \item{land}{elevation [m]}
  \item{area}{PixelArea_km2 [m]}
  \item{AreaOfCoastalZone}{Surface of Sea floor with a depth of between 0 and 200 meters [km2]}
  etc...
}
\description{
Calculate yearly means for event metrics.
}
\details{
This functions queries the FTP server where the NetCDF files are stored. It will
retrieve the one file that matches the name provided to it via the \code{fjord}
argument. Note that these files can be multiple gigabytes in size.
}
\examples{
fl_DownloadFjord("kong", dirdata = "data/PAR")

}
\author{
Bernard Gentili, Robert W. Schlegel
}
