% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fl_ListFjords.R
\name{fl_ListFjords}
\alias{fl_ListFjords}
\title{Simply lists the names of fjords with available data.}
\usage{
fl_ListFjords()
}
\value{
A list of currently 8 different character vectors representing a range of fjords
in the EU Arctic.
}
\description{
Run this to determine which character vectors to use when downloading data
via \code{\link{fl_DownloadFjord}}, or a range of other uses within \code{FjordLight}.
}
\examples{
fl_ListFjords()

}
\author{
Bernard Gentili
}
