% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flplot_bathymetry.R
\name{flplot_bathymetry}
\alias{flplot_bathymetry}
\title{Plot bathymetry data found within NetCDF files downloaded by \code{FjordLight}.}
\usage{
flplot_bathymetry(r, l = NULL, name)
}
\arguments{
\item{r}{The bathymetry data in a raster format.}

\item{l}{The land data in the same format.}

\item{name}{The name of the chosen fjord.}
}
\value{
A base R plot of the bathymetric data will be provided.
}
\description{
Internal function used by \code{\link{flget_bathymetry}} that plots the bathymetry data
contained within the NetCDF files downloaded via \code{\link{fl_DownloadFjord}}.
}
\author{
Bernard Gentili
}
\keyword{internal}
