% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flplot_climatology.R
\name{flplot_climatology}
\alias{flplot_climatology}
\title{Plot annual/monthly climatology data extracted from NetCDF files downloaded by \code{FjordLight}.}
\usage{
flplot_climatology(r, l, name, optics, period, month, year)
}
\arguments{
\item{r}{The yearly/monthly climatology data in a raster format.}

\item{l}{The land data in the same format.}

\item{name}{The name of the chosen fjord.}

\item{optics}{The name of the PAR variable being displayed.}

\item{period}{The period being displayed. I.e. yearly or monthly.}

\item{month}{If monthly climatology data, which month(s) is being displayed.}

\item{year}{If yearly data, which year(s) is being displayed}
}
\value{
A base R plot of the yearly or monthly climatology data will be provided.
}
\description{
Internal function used by \code{\link{flget_climatology}} that plots data
contained within the NetCDF files downloaded via \code{\link{fl_DownloadFjord}}.
}
\author{
Bernard Gentili
}
\keyword{internal}
