% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flplot_KPARMonthlyTS.R
\name{flplot_KPARMonthlyTS}
\alias{flplot_KPARMonthlyTS}
\title{Plot monthly K_PAR data extracted from NetCDF files downloaded by \code{FjordLight}.}
\usage{
flplot_KPARMonthlyTS(r)
}
\arguments{
\item{r}{The monthly K_PAR data in a raster stack format.}
}
\value{
A base R plot of the monthly K_PAR data will be provided.
}
\description{
Internal function used by \code{\link{flget_KPARMonthlyTS}} that plots data
contained within the NetCDF files downloaded via \code{\link{fl_DownloadFjord}}.
}
\author{
Bernard Gentili
}
\keyword{internal}
