% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flplot_PARbottomMonthlyTS.R
\name{flplot_PARbottomMonthlyTS}
\alias{flplot_PARbottomMonthlyTS}
\title{Plot monthly bottom PAR data extracted from NetCDF files downloaded by \code{FjordLight}.}
\usage{
flplot_PARbottomMonthlyTS(r)
}
\arguments{
\item{r}{The monthly bottom PAR data in a raster stack format.}
}
\value{
A base R plot of the monthly bottom PAR data will be provided.
}
\description{
Internal function used by \code{\link{flget_PARbottomMonthlyTS}} that plots data
contained within the NetCDF files downloaded via \code{\link{fl_DownloadFjord}}.
}
\author{
Bernard Gentili
}
\keyword{internal}
