% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flget_bathymetry.R
\name{flget_bathymetry}
\alias{flget_bathymetry}
\title{Extract and/or plot bathymetry data from a NetCDF file downloaded by \code{FjordLight}.}
\usage{
flget_bathymetry(fjord, what = "o", mode = "raster", PLOT = FALSE)
}
\arguments{
\item{fjord}{Expects the object loaded via \code{\link{fl_LoadFjord}}.}

\item{what}{The default value \code{"s"} will load all "sea" data, meaning it will filter
out any land pixels. The other options are: \code{"c"} filters out only bathymetry data
at a depth of 200 m to 0 m, and \code{"sl"} loads the entire data layer (all sea and land pixels).}

\item{mode}{Determines the format of the bathymetry data loaded into the R environment.
The default \code{"raster"} will load the data as a raster format. The other option \code{"3col"}
will load the data as a data.frame with three columns.}

\item{PLOT}{Boolean argument (default = \code{FALSE}) that tells the function if the user
would like the loaded bathymetry data to be plotted or not.}
}
\value{
Depending on which arguments the user chooses, this function will return the
filtered bathymetry data as a \code{RasterLayer} (\code{mode = "raster"}) or
data.frame (\code{mode = "3col"}). The data.frame will contain the following columns:
  \item{longitude}{degree decimals}
  \item{latitude}{degree decimals}
  \item{depth}{metres}
}
\description{
This functions will conveniently extract the bathymetry data stored within a
NetCDF file downloaded via \code{\link{fl_DownloadFjord}}. There are options
for how the user would like to subset the data, which data format the data
should be extracted to, and if the user would like to plot the data in the process.
}
\examples{
\dontrun{
# Download+load data
fjord <- "kong"
fl_DownloadFjord(fjord, dirdata = "data/PAR")
fjorddata <- fl_LoadFjord(fjord, dirdata = "data/PAR")

# all depths (what = "s" ; s for Sea), as raster
bathy <- flget_bathymetry(fjorddata, what = "s", mode = "raster", PLOT = TRUE)
# coastal zone [0-200m] (what = "c" ; c for Coastal), as raster
bathy <- flget_bathymetry(fjorddata, what = "c", mode = "raster", PLOT = TRUE)
# as 3 columns data frame (mode = "3col" : longitude, latitude, depth)
sea <- flget_bathymetry(fjorddata, what = "s", mode = "3col", PLOT = FALSE)
cz <- flget_bathymetry(fjorddata, what = "c", mode = "3col", PLOT = FALSE)
# you may add letter "l" if you want land elevation
sealand <- flget_bathymetry(fjorddata, what = "sl", mode = "3col", PLOT = FALSE)
}

}
\author{
Bernard Gentili
}
